extracted label-scheme to its own class
