+++ src/app/data-loader/info/workspace-info.ts
@@ -1,16 +1,14 @@
+import { LabelScheme } from './label-scheme';
 import { VideoInfo } from './video-info';
-
 import { DataInfo } from './data-info';
 
-import { invert } from '../../util/util';
-import { Synchronizer } from '../../util/sync';
 
 // #region [Interfaces]
 export interface TypeMap {
     [index: number]: string;
 }
 
-type LabelKey = string |number
+export type LabelKey = string |number
 type Info = DataInfo | LabelScheme
 // #endregion
 
@@ -96,55 +94,5 @@ export class WorkspaceInfo {
         }
         return obj;
     }
-    // #endregion
-}
-
-export class LabelScheme {
-    // #region [Constants]
-    static NULL_EVENT = 'Ø';
-    static NULL_KEY = 0;
-    // #endregion
-
-    // #region [Properties]
-    workspace: string;
-    name: string;
-    event_map: TypeMap;
-    null_label: LabelKey;
-    path?: string;
-    video?: string;
-    private ws: WorkspaceInfo;
-    // #endregion
-
-    // #region [Constructor]
-    constructor(ws, info) {
-        this.ws = ws;
-        this.workspace = info.workspace || ws.name;
-        this.name = info.name;
-        this.event_map = info.event_map;
-        if ("path" in info) this.path = info.path;
-        if ("video" in info) this.video = info.video;
-        this.null_label = info.null_label || LabelScheme.NULL_KEY;
-    }
-    // #endregion
-
-    // #region [Accessors]
-    get flashes() {
-        if (!this.video) return [];
-        return this.ws.vFlashes(this.video);
-    }
-
-    get hasLabels() { return !!this.path }
-
-    get inv_event_map() { return invert(this.event_map) }
-    // #endregion
-
-    // #region [Public Methods]
-    sync(dataset: string): Synchronizer {
-        let ds = this.ws.getData(dataset);
-        if (!ds) return null;
-        return new Synchronizer(ds.flashes, this.flashes);
-    }
-
-    lblKey(type: string) { return parseInt(this.inv_event_map[type]) }
     // #endregion
 }
