+++ src/app/home/workspace/workspace.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { WorkspaceInfo, LabelScheme } from '../../data-loader/info/workspace-info';
+import { WorkspaceInfo } from '../../data-loader/info/workspace-info';
+import { LabelScheme } from "../../data-loader/info/label-scheme";
 import { DataInfo } from "../../data-loader/info/data-info";
 
 // #region [Interfaces]
