+++ src/app/saver/saver.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { TypeMap, LabelScheme } from '../data-loader/info/workspace-info';
+import { TypeMap } from '../data-loader/info/workspace-info';
+import { LabelScheme } from "../data-loader/info/label-scheme";
 import { VideoInfo } from "../data-loader/info/video-info";
 import { DataInfo } from "../data-loader/info/data-info";
 import { Label } from '../dataview/labelstreams/labelstream';
